optimisation des controllers et views chauffeur, backoffice client,admin,chauffeurs

+++ database/migrations/2022_02_25_215332_create_users_table.php
@@ -17,7 +17,7 @@ class CreateUsersTable extends Migration
             $table->id();
             $table->string('nom');
             $table->string('prenom');
-            $table->float('numero_telephone');
+            $table->integer('numero_telephone');
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
