+++ database/migrations/2022_02_26_070315_create_vehicules_table.php
@@ -33,7 +33,7 @@ class CreateVehiculesTable extends Migration
             ->onUpdate('cascade');
             $table->boolean('statut_vehicule');
             $table->string('immatriculation')->unique();
-            $table->float('kilometrage');
+            $table->integer('kilometrage');
             $table->string('numero_chassi')->unique();
             $table->date('date_debut_assurance');
             $table->date('date_fin_assurance');
